test project config

+++ src/services/RelabelService.php
@@ -20,7 +20,10 @@ use craft\base\Component;
 use Craft;
 use craft\elements\Entry;
 use craft\elements\User;
+use craft\events\ConfigEvent;
+use craft\helpers\Db;
 use craft\helpers\Json;
+use craft\helpers\StringHelper;
 use craft\models\FieldLayout;
 
 /**
@@ -34,11 +37,12 @@ use craft\models\FieldLayout;
  */
 class RelabelService extends Component
 {
-
+    const CONFIG_RELABEL_KEY = 'relabel';
     /**
      * Event to register a field layout ID for custom elements
      */
     const EVENT_REGISTER_LABELS = 'eventRegisterLabels';
+
     /**
      * @return RelabelRecord[]
      */
@@ -64,10 +68,10 @@ class RelabelService extends Component
                 'fields.handle',
                 'fields.name as oldName'
             ]
-        )->from('{{%relabel}} as relabel')
-            ->where(['fieldLayoutId' => $layoutId])
-            ->leftJoin('{{%fields}} as fields', '[[fields.id]] = [[relabel.fieldId]]')
-            ->all();
+        )->from('{{%relabel}} as relabel')->where(['fieldLayoutId' => $layoutId])->leftJoin(
+                '{{%fields}} as fields',
+                '[[fields.id]] = [[relabel.fieldId]]'
+            )->all();
     }
 
     /**
@@ -76,14 +80,15 @@ class RelabelService extends Component
      * @return \craft\models\FieldLayout|null
      * @throws \yii\base\InvalidConfigException
      */
-    public function getLayoutFromRequest(){
+    public function getLayoutFromRequest()
+    {
         $request = Craft::$app->getRequest();
         $segments = $request->segments;
         $layout = null;
         if (\count($segments) >= 1) {
             switch ($segments[0]) {
                 case 'entries':
-                    if(\count($segments) <= 1){
+                    if (\count($segments) <= 1) {
                         return null;
                     }
                     $lastSegment = $segments[\count($segments) - 1];
@@ -103,7 +108,7 @@ class RelabelService extends Component
 
                     break;
                 case 'categories':
-                    if(\count($segments) <= 1){
+                    if (\count($segments) <= 1) {
                         return null;
                     }
                     if ($groupHandle = $segments[1]) {
@@ -113,7 +118,7 @@ class RelabelService extends Component
                     }
                     break;
                 case 'globals':
-                    if(\count($segments) <= 1){
+                    if (\count($segments) <= 1) {
                         return null;
                     }
                     $handle = $segments[\count($segments) - 1];
@@ -128,67 +133,71 @@ class RelabelService extends Component
                     $layout = Craft::$app->getFields()->getLayoutByType(User::class);
                     break;
                 case 'gift-voucher':
-                    if(\count($segments) <= 2){
+                    if (\count($segments) <= 2) {
                         return null;
                     }
 
                     // check for an id
-                    if(\count($segments) === 4){
-                        if(\is_numeric($segments[3])){
-                            $element = Craft::$app->getElements()->getElementById((int)$segments[3], 'verbb\\giftvoucher\\elements\\Voucher');
+                    if (\count($segments) === 4) {
+                        if (\is_numeric($segments[3])) {
+                            $element = Craft::$app->getElements()->getElementById(
+                                (int) $segments[3],
+                                'verbb\\giftvoucher\\elements\\Voucher'
+                            );
                             $layout = $element->getFieldLayout();
-
-                        }else{
+                        } else {
                             // unfortunately we can't just use Commerce classes since we can't make
                             // sure they exists and the plugin should run even without it :(
                             $type = $segments[2];
-                            $fieldLayoutId = (new Query())
-                                ->select(['fieldLayoutId'])
-                                ->from('{{%giftvoucher_vouchertypes}}')
-                                ->where(['handle' => $type])
-                                ->scalar();
+                            $fieldLayoutId = (new Query())->select(['fieldLayoutId'])->from(
+                                    '{{%giftvoucher_vouchertypes}}'
+                                )->where(['handle' => $type])->scalar();
                             /** @var \craft\models\Section $section */
-                            if ($fieldLayoutId !== false){
-                                $layout = Craft::$app->getFields()->getLayoutById((int)$fieldLayoutId);
+                            if ($fieldLayoutId !== false) {
+                                $layout = Craft::$app->getFields()->getLayoutById((int) $fieldLayoutId);
                             }
                         }
                     }
 
                     break;
                 case 'commerce':
-                    if(\count($segments) <= 2){
+                    if (\count($segments) <= 2) {
                         return null;
                     }
 
-                    if($segments[1] === 'orders'){
+                    if ($segments[1] === 'orders') {
                         $lastSegment = $segments[\count($segments) - 1];
-                        if(\is_numeric($lastSegment)){
-                            $element = Craft::$app->getElements()->getElementById($lastSegment, 'craft\\commerce\\elements\\Order');
+                        if (\is_numeric($lastSegment)) {
+                            $element = Craft::$app->getElements()->getElementById(
+                                $lastSegment,
+                                'craft\\commerce\\elements\\Order'
+                            );
                             $layout = $element->getFieldLayout();
                         }
                     }
 
-                    if($segments[1] === 'products'){
+                    if ($segments[1] === 'products') {
                         // unfortunately we can't just use Commerce classes since we can't make
                         // sure they exists and the plugin should run even without it :(
                         $lastSegment = $segments[\count($segments) - 1];
                         $id = explode('-', $lastSegment)[0];
                         if ($id && strpos($lastSegment, '-')) {
                             /** @var Element $element */
-                            $element = Craft::$app->getElements()->getElementById($id, 'craft\\commerce\\elements\\Product');
+                            $element = Craft::$app->getElements()->getElementById(
+                                $id,
+                                'craft\\commerce\\elements\\Product'
+                            );
                             $layout = $element->getFieldLayout();
                         } else {
-                            $productGroup =  $segments[2];
+                            $productGroup = $segments[2];
                             // query for it
-                            $fieldLayoutId = (new Query())
-                                ->select(['fieldLayoutId'])
-                                ->from('{{%commerce_producttypes}}')
-                                ->where(['handle' => $productGroup])
-                                ->scalar();
+                            $fieldLayoutId = (new Query())->select(['fieldLayoutId'])->from(
+                                    '{{%commerce_producttypes}}'
+                                )->where(['handle' => $productGroup])->scalar();
 
                             /** @var \craft\models\Section $section */
-                            if ($fieldLayoutId !== false){
-                                $layout = Craft::$app->getFields()->getLayoutById((int)$fieldLayoutId);
+                            if ($fieldLayoutId !== false) {
+                                $layout = Craft::$app->getFields()->getLayoutById((int) $fieldLayoutId);
                             }
                         }
                     }
@@ -220,22 +229,24 @@ class RelabelService extends Component
             $attributes = $request->getBodyParam('attributes');
             $elementId = $request->getBodyParam('elementId');
             $elementType = $request->getBodyParam('elementType');
-            $siteId = (int)$request->getBodyParam('siteId');
+            $siteId = (int) $request->getBodyParam('siteId');
             if ($elementId) {
-                $element = Craft::$app->getElements()->getElementById((int)$elementId, $elementType, $siteId);
+                $element = Craft::$app->getElements()->getElementById((int) $elementId, $elementType, $siteId);
             } else {
                 $element = new $elementType();
                 Craft::configure($element, $attributes);
             }
 
-            if($element !== null && $element::hasContent()){
+            if ($element !== null && $element::hasContent()) {
                 $layout = $element->getFieldLayout();
             }
         }
 
-        $event = new RegisterLabelEvent([
-                                            'fieldLayoutId' => $layout !== null? $layout->id : null
-                                        ]);
+        $event = new RegisterLabelEvent(
+            [
+                'fieldLayoutId' => $layout !== null ? $layout->id : null
+            ]
+        );
         $this->trigger(self::EVENT_REGISTER_LABELS, $event);
 
         if ($event->fieldLayoutId !== null) {
@@ -267,13 +278,15 @@ class RelabelService extends Component
         $layout = $this->getLayoutFromRequest();
 
         // fire an event, so others may include custom labels
-        $event = new RegisterLabelEvent([
-                                            'fieldLayoutId' => $layout !== null? $layout->id : null
-                                        ]);
+        $event = new RegisterLabelEvent(
+            [
+                'fieldLayoutId' => $layout !== null ? $layout->id : null
+            ]
+        );
         $this->trigger(self::EVENT_REGISTER_LABELS, $event);
 
         // if there is a field layout, grab new labels
-        if($event->fieldLayoutId !== null){
+        if ($event->fieldLayoutId !== null) {
             $labelsForLayout = $this->getAllLabelsForLayout($event->fieldLayoutId);
         }
 
@@ -291,16 +304,19 @@ class RelabelService extends Component
     }
 
     /**
-     * @param $layout
-     * @param $relabel
-     * @throws \Exception
-     * @throws \Throwable
-     * @throws \yii\db\StaleObjectException
+     * @param \craft\models\FieldLayout $layout
+     * @param                           $relabel
+     *
+     * @throws \yii\base\ErrorException
+     * @throws \yii\base\Exception
+     * @throws \yii\base\NotSupportedException
+     * @throws \yii\web\ServerErrorHttpException
      */
     public function saveRelabelsForLayout(FieldLayout $layout, $relabel)
     {
         if ($relabel !== null && \is_array($relabel)) {
             foreach ($relabel as $fieldId => $values) {
+                /** @var RelabelRecord $record */
                 $record = RelabelRecord::find()->where(
                     [
                         'fieldId'       => $fieldId,
@@ -308,21 +324,42 @@ class RelabelService extends Component
                     ]
                 )->one();
 
-                if ($record && !$values['name'] && !$values['instructions']) {
-                    $record->delete();
+                $isNew = $record === null;
+
+                if ($isNew === false && !$values['name'] && !$values['instructions']) {
+                    // remove it
+                    $path = self::CONFIG_RELABEL_KEY . '.' . $record->uid;
+                    Craft::$app->getProjectConfig()->remove($path);
+                    //$record->delete();
                     continue;
                 }
 
                 if ($record === null) {
                     $record = new RelabelRecord();
+                    $record->uid = StringHelper::UUID();
+                } else if (!$record->uid) {
+                    $record->uid = Db::uidById('{{%relabel}}', $record->id);
                 }
+
                 $record->fieldId = $fieldId;
                 $record->name = $values['name'];
                 $record->instructions = $values['instructions'];
                 $record->fieldLayoutId = $layout->id;
-                if (!$record->save()) {
-                    Craft::error('[Relabel] could not store field layout ' . Json::encode($record->getErrors()), Relabel::class);
-                }
+                $path = self::CONFIG_RELABEL_KEY . '.' . $record->uid;
+
+                Craft::$app->getProjectConfig()->set(
+                    $path,
+                    [
+                        'field'        => Db::uidById('{{%fields}}', (int) $fieldId),
+                        'fieldLayout'  => $layout->uid,
+                        'instructions' => $values['instructions'],
+                        'name'         => $values['name']
+                    ]
+                );
+
+                //if (!$record->save()) {
+                //    Craft::error('[Relabel] could not store field layout ' . Json::encode($record->getErrors()), Relabel::class);
+                //}
             }
         }
 
@@ -332,10 +369,66 @@ class RelabelService extends Component
             ['fieldLayoutId' => $layout->id]
         )->all();
         foreach ($unusedLabels as $record) {
-            $record->delete();
+            $path = self::CONFIG_RELABEL_KEY . '.' . $record->uid;
+            Craft::$app->getProjectConfig()->remove($path);
         }
     }
 
+    /**
+     * @param \craft\events\ConfigEvent $event
+     *
+     * @throws \Throwable
+     * @throws \yii\db\StaleObjectException
+     * since 28.01.2019
+     */
+    public function handleDeletedRelabel(ConfigEvent $event)
+    {
+        // Get the UID that was matched in the config path
+        $uid = $event->tokenMatches[0];
+
+        // Get the product type
+        /** @var RelabelRecord $record */
+        $record = RelabelRecord::findOne(['uid' => $uid]);
+
+        // If that came back empty, we're done!
+        if ($record === null) {
+            return;
+        }
+
+        $record->delete();
+    }
+
+    /**
+     * @param \craft\events\ConfigEvent $event
+     * since 28.01.2019
+     */
+    public function handleChangedRelabel(ConfigEvent $event)
+    {
+        // Get the UID that was matched in the config path
+        $uid = $event->tokenMatches[0];
+
+
+        $record = RelabelRecord::findOne(['uid' => $uid]);
+
+        if ($record === null) {
+            $record = new RelabelRecord();
+        }
+
+        $record->uid = $uid;
+        $record->fieldId = Db::idByUid('{{%fields}}', $event->newValue['field']);
+        $record->fieldLayoutId = Db::idByUid('{{%fieldlayouts}}', $event->newValue['fieldLayout']);
+        $record->instructions = $event->newValue['instructions'];
+        $record->name = $event->newValue['name'];
+
+        $record->save();
+    }
+
+    public function deleteRelabel($relabel)
+    {
+        // Remove it from the project config
+        $path = self::CONFIG_RELABEL_KEY . ".{$relabel->uid}";
+        Craft::$app->projectConfig->remove($path);
+    }
 
     /**
      * @return FieldLayout|null
@@ -343,14 +436,13 @@ class RelabelService extends Component
     public function getLayoutByTypeId()
     {
         $typeId = Craft::$app->getRequest()->getBodyParam('typeId');
-        $fieldLayoutId = (new Query())
-            ->select(['fieldLayoutId'])
+        $fieldLayoutId = (new Query())->select(['fieldLayoutId'])
             ->from('{{%entrytypes}}')
             ->where(['id' => $typeId])
             ->scalar();
         $layout = null;
-        if($fieldLayoutId !== null && $fieldLayoutId !== false){
-            $layout = Craft::$app->getFields()->getLayoutById((int)$fieldLayoutId);
+        if ($fieldLayoutId !== null && $fieldLayoutId !== false) {
+            $layout = Craft::$app->getFields()->getLayoutById((int) $fieldLayoutId);
         }
 
         return $layout;
